+++ lib/ajax-datatables-rails/error.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module AjaxDatatablesRails
+  module Error
+    class BaseError < StandardError; end
+    class InvalidSearchCondition < BaseError; end
+  end
+end